rust-server-cleanup(scheduler-split,non_mechanical_provable): split scheduler intake helpers
