+++ rust/sglang-server/src/tokenizer_manager.rs
@@ -4,5 +4,7 @@ pub mod channel;
 pub mod detokenizer;
 pub mod from_scheduler;
 pub mod to_scheduler;
+mod to_scheduler_types;
+mod to_scheduler_validation;
 pub mod tokenizer;
 pub mod wiring;
